made type definition of StrintRef more compatible with older compilers by using c_size_t instead of c_int32_t. BTW: for storing a string length c_size_t makes more sense anyways!
git-svn-id: https://subversion.assembla.com/svn/zappralot/trunk/libfortres@350 c37337d0-301c-4318-a67e-7f75229f9ee7

+++ string_ref.f90
@@ -4,8 +4,8 @@ module string_ref
   implicit none
 
   type, public, bind(C) :: StringRef
-    type (c_ptr)            :: loc = c_null_ptr
-    integer(kind=c_int32_t) :: len = 0
+    type (c_ptr)           :: loc = c_null_ptr
+    integer(kind=c_size_t) :: len = 0
   end type
 
   interface str
